fix: 기존 viewController에서 view를 그리는 코드를 cardGameView로 이동, add : UISegmentedControl 설정 추가
